+++ services/web/server/src/simcore_service_webserver/diagnostics/_healthcheck.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass, field
 from aiohttp import web
 from servicelib.aiohttp.incidents import LimitedOrderedStack, SlowCallback
 
-from ..rest.rest_healthcheck import HealthCheckFailed
+from ..rest.healthcheck import HealthCheckFailed
 from .settings import get_plugin_settings
 
 _logger = logging.getLogger(__name__)
