+++ services/web/server/src/simcore_service_webserver/rest/plugin.py
@@ -20,7 +20,7 @@ from .._constants import APP_OPENAPI_SPECS_KEY, APP_SETTINGS_KEY
 from .._meta import API_VTAG, api_version_prefix
 from ..security.plugin import setup_security
 from . import _handlers
-from .rest_healthcheck import HealthCheck
+from .healthcheck import HealthCheck
 from .rest_settings import RestSettings, get_plugin_settings
 from .rest_utils import get_openapi_specs_path, load_openapi_specs
